refactor(#248): FSD 파일 구조 수정
