+++ src/entities/dashboard/api/dashboard.api.ts
@@ -1,6 +1,6 @@
 import { apiClient } from "@/shared/lib/api/client";
 
-import { ProgressSuccessResponse } from "./types";
+import { ProgressSuccessResponse } from "../types/dashboard.types";
 
 export const progressApi = {
   read: () => apiClient.get<ProgressSuccessResponse>("/api/main/progress"),
