+++ src/entities/dashboard/index.ts
@@ -1,0 +1,8 @@
+export { progressApi } from "./api/dashboard.api";
+export type {
+  ProgressData,
+  ProgressErrorResponse,
+  ProgressItem,
+  ProgressResponse,
+  ProgressSuccessResponse,
+} from "./types/dashboard.types";
