+++ src/entities/dashboard/types/dashboard.types.ts
