+++ src/features/notification/notification.utils.ts
@@ -51,44 +51,3 @@ export const buildNotificationUrl = ({ goalId, teamId }: BuildUrlParams) => {
 
   return `/taskmate/personal/goal/${goalId}`;
 };
-
-// 무한스크롤 옵션 객체 생성 유틸 함수
-// todo: main utils에 중복 정의 하나의 공통유틸로 분리
-
-type CursorParams = {
-  size?: number;
-  cursorId?: number;
-  cursorCreatedAt?: string;
-};
-
-type CursorPage = {
-  hasNext: boolean;
-  nextCursorId?: number;
-  nextCursorCreatedAt?: string;
-};
-
-export function createPaginationOptions<
-  Params extends CursorParams = CursorParams,
-  Page extends CursorPage = CursorPage,
->(apiKey: string, apiFunction: (param: Params) => Promise<{ data: Page }>) {
-  const size = 20;
-
-  return {
-    queryKey: [apiKey],
-    initialPageParam: { size } as Params,
-
-    queryFn: async ({ pageParam }: { pageParam: Params }) => {
-      const res = await apiFunction(pageParam);
-      return res.data;
-    },
-
-    getNextPageParam: (lastPage: Page): Params | undefined =>
-      lastPage.hasNext
-        ? ({
-            size,
-            cursorId: lastPage.nextCursorId,
-            cursorCreatedAt: lastPage.nextCursorCreatedAt,
-          } as Params)
-        : undefined,
-  };
-}
