+++ src/widgets/home/FavoriteGoalsSection/FavoriteGoalsItem.tsx
